move prepareRedirects to scripts folder
